requirements.txt
django~=4.2.7
django-leaflet~=0.29.0
djangorestframework~=3.14.0
geopy~=2.4.0
whitenoise~=6.6.0